Create useActiveDescendant hook
